+++ src/components/general/character/Poses.jsx
@@ -1,13 +1,40 @@
-
-import React, { useEffect, useState, useRef, useContext } from 'react';
+import React, {useState, useRef} from 'react';
 import classnames from 'classnames';
 
 import styles from './poses.module.css';
 
-const poses = [
-    'dance',
-    'happy',
-];
+import metaversefile from 'metaversefile';
+import {emoteAnimations} from '../../../../avatars/animationHelpers.js';
+import emotes from './emotes.json';
+
+//
+
+let emoteTimeout = null;
+export const triggerEmote = emote => {
+  // clear old emote
+  const localPlayer = metaversefile.useLocalPlayer();
+  localPlayer.removeAction('emote');
+  if (emoteTimeout) {
+    clearTimeout(emoteTimeout);
+    emoteTimeout = null;
+  }
+
+  // add new emote
+  const newAction = {
+    type: 'emote',
+    animation: emote,
+  };
+  localPlayer.addAction(newAction);
+
+  const emoteAnimation = emoteAnimations[emote];
+  const emoteAnimationDuration = emoteAnimation.duration;
+  emoteTimeout = setTimeout(() => {
+    const actionIndex = localPlayer.findActionIndex(action => action.type === 'emote' && action.animation === emote);
+    localPlayer.removeActionIndex(actionIndex);
+    
+    emoteTimeout = null;
+  }, emoteAnimationDuration * 1000);
+};
 
 //
 
@@ -16,6 +43,10 @@ export const Poses = ({
 }) => {
     const [ posesOpen, setPosesOpen ] = useState( false );
     const posesRef = useRef();
+
+    const poseClick = emote => e => {
+      triggerEmote(emote);
+    };
 
     return (
         <div
@@ -32,14 +63,16 @@ export const Poses = ({
             }}
             ref={posesRef}
         >
-            {poses.map((pose, poseIndex) => {
+            {emotes.map((emote, emoteIndex) => {
+                const {name, icon} = emote;
                 return (
                     <div
                         className={styles.pose}
-                        key={pose}
+                        onClick={poseClick(name)}
+                        key={name}
                     >
-                        <img src={`images/poses/${pose}.svg`} className={styles.poseIcon} />
-                        <div className={styles.poseName}>{pose}</div>
+                        <img src={`images/poses/${icon}`} className={styles.poseIcon} />
+                        <div className={styles.poseName}>{name}</div>
                     </div>
                 );
             })}
