+++ src/components/general/character/emotions.module.css
@@ -107,7 +107,7 @@
   display: flex;
   position: relative;
   align-items: center;
-  width: 90px;
+  width: 120px;
   padding: 0 10px;
   text-align: center;
   color: #666;
