+++ src/components/general/character/poses.module.css
@@ -15,8 +15,7 @@
 .poses:not(.open) .pose:first-child {
   display: flex;
 }
-.poses:not(.open) .pose:first-child .poseName,
-.poses:not(.open) .pose:first-child .poseProgress
+.poses:not(.open) .pose:first-child .poseName
 {
   display: none;
 }
@@ -33,67 +32,21 @@
   user-select: none;
   transition: all 1s cubic-bezier(0, 1, 0, 1);
 }
-.poses .pose .poseProgress,
 .poses .pose .poseName,
-.poses .pose .poseIcon,
-.poses .pose .poseNamePlaceholder,
-.poses .pose .poseIconPlaceholder
+.poses .pose .poseIcon
 {
   pointer-events: none;  
-}
-.poses .pose .poseProgress {
-  position: absolute;
-  width: 50px;
-  height: 140px;
-  left: -50px;
-  -webkit-appearance: none;
-  transform: rotateZ(-90deg);
-  transition: color 1s cubic-bezier(0, 1, 0, 1);
-  transform-origin: 100% 0;
-}
-.poses .pose .poseProgress::-webkit-progress-bar {
-  background: transparent;
-}
-.poses .pose .poseProgress::-webkit-progress-value {
-  background-color: #444;
-  /* background-image: linear-gradient(to bottom, #56ab2f, #a8e063); */
-  transition: background-color 1s cubic-bezier(0, 1, 0, 1);
 }
 .poses .pose:active {
   filter: invert(1);
 }
-/* .poses .pose.full .poseProgress::-webkit-progress-value {
-  background-color: #000;
-} */
 .poses .pose .poseIcon {
   width: 50px;
   height: 50px;
+  padding: 5px;
   object-fit: contain;
   filter: brightness(0.3);
   /* transition: filter 1s cubic-bezier(0, 1, 0, 1); */
-}
-
-.poses .pose .poseIconPlaceholder {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 50px;
-  height: 50px;
-  /* background-color: #333; */
-}
-.poses .pose .poseNamePlaceholder {
-  position: absolute;
-  top: 0;
-  left: 50px;
-  width: calc(140px - 50px);
-  height: 50px;
-  /* background-color: #111; */
-}
-.poses .pose:hover .poseIconPlaceholder {
-  background-color: #222;
-}
-.poses .pose:hover .poseNamePlaceholder {
-  background-color: #111;
 }
 
 .poses .pose:hover .poseIcon,
@@ -108,7 +61,7 @@
   display: flex;
   position: relative;
   align-items: center;
-  width: 90px;
+  width: 120px;
   padding: 0 10px;
   text-align: center;
   color: #666;
